requirement.txt
absl-py==0.9.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cloudpickle==1.4.1
cycler==0.10.0
Cython==0.29.19
decorator==4.4.2
defusedxml==0.6.0
detectron2==0.1.3+cu101
entrypoints==0.3
future==0.18.2
fvcore==0.1.1.post20200608
google-auth==1.16.1
google-auth-oauthlib==0.4.1
grpcio==1.29.0
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mock==4.0.2
nb-conda-kernels==2.2.3
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1591485221441/work
oauthlib==3.1.0
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
portalocker==1.7.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools @ git+https://github.com/cocodataset/cocoapi.git@8c9bcc3cf640524c4c20a9c40e89cb6a2f2fa0e9#subdirectory=PythonAPI
pydot==1.4.1
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-git==2018.2.1
pytz==2020.1
PyYAML==5.1
pyzmq==18.1.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.15.0
tabulate==0.8.7
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.5.0+cu101
torchvision==0.6.0+cu101
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
yacs==0.1.7
zipp==3.1.0
